합성 완료

+++ Assets/Scripts/Game/GameStructs/Status.cs
@@ -1,10 +1,10 @@
 public struct Status 
 {
     //물리 공격력
-    public float AD;
+    public int AD;
     
     //마법 공격력
-    public float AP;
+    public int AP;
 
     //공격 속도
     public float A_SPD;
@@ -16,7 +16,7 @@ public struct Status
     public float M_SPD;
     
     //방어력
-    public float DEF;
+    public int DEF;
     
     //최대 체력
     public int MAX_HP;
